#리팩토링 front 폴더 구조 변경 및 컴포넌트 분리
