+++ back/front/src/pages/Main/Home/graphql/schema.ts
@@ -1,0 +1,23 @@
+import { gql } from "@apollo/client";
+
+export const PRODUCTS = gql`
+  query Products {
+    products {
+      sId
+      sName
+      sTitle
+      sContents
+      sPrice
+      sView
+      slideImg
+      mainBottomImg
+    }
+  }
+`;
+export const COUNT_VIEW = gql`
+  mutation CountView($CountViewId: Int) {
+    countView(id: $countViewId) {
+      sId
+    }
+  }
+`;
