+++ back/front/src/pages/Main/Home/hooks/queries/useGetProducts.ts
@@ -1,0 +1,11 @@
+import { useQuery } from "@apollo/client";
+import { Products } from "../../../../../types/Products.interface";
+import { PRODUCTS } from "../../graphql/schema";
+
+function useGetProducts() {
+  const { data, loading } = useQuery<Products>(PRODUCTS);
+
+  return { data, loading };
+}
+
+export default useGetProducts;
