+++ back/front/src/types/Cart.interface.ts
@@ -1,0 +1,12 @@
+export interface Cart {
+  bId: number;
+  sId: number;
+  user_code: number;
+  bDate: string;
+  sName: string;
+  sPrice: number;
+  bCount: number;
+  slideImg: [string];
+  cartId: number;
+  check: boolean;
+}
