+++ back/front/src/types/User.interface.ts
@@ -1,0 +1,8 @@
+export interface User {
+  user_code: number;
+  kakao_id: string;
+  kakao_profile_img: string;
+  kakao_nickname: string;
+  kakao_email: string;
+  user_role: string;
+}
